Major bus card update
Updates are now differed and no longer refresh in seconds, the reason is primarily for performance. It becomes incredibly taxing to do as such as we have to constantly recreate cards per second.

There are still issues with this method and looking into solving them soon.

Issue 1. Cards when scrolled away and back may be missing bus details or even displayed incorrectly. The main issue is function updateUI calling every card notifyItemChanged. If Major update is false, it will not render details.